+++ finagle-http/src/main/scala/com/twitter/finagle/http/defaultClientProtocol.scala
@@ -1,0 +1,10 @@
+package com.twitter.finagle.http
+
+import com.twitter.app.GlobalFlag
+
+/**
+ * [[GlobalFlag]] that can be used to configure the default protocol used when creating a new
+ * [[com.twitter.finagle.Http.Client]].
+ */
+object defaultClientProtocol
+    extends GlobalFlag[Protocol](Protocol.Default, "Default HTTP Client Protocol")