+++ finagle-http/src/test/scala/com/twitter/finagle/http/filter/ClientNackFilterTest.scala
@@ -103,8 +103,10 @@ class ClientNackFilterTest extends AnyFunSuite {
 
       assert(await(client(request)).status == http.Status.Ok)
 
-      // Should have closed the connection on the first nack
-      assert(clientSr.counters(Seq("http", "connects")) == 2)
+      // h2 sessions should remain at 1 and the connection should not sever
+      assert(clientSr.counters(Seq("http", "connects")) == 1)
+      assert(clientSr.gauges(Seq("http", "h2pool-sessions"))() == 1.0f)
+      assert(clientSr.counters(Seq("http", "failures", "rejected")) == 1)
       assert(serverSr.counters(Seq("myservice", "nacks")) == 1)
 
       closeCtx()